Extract workspace render state hook
